DOC/TST: revised indexing section in docs          updated whatsnew          all tests work
DOC: changes suggested by Jan Schulz
     revised whatsnew to include mostly references to new indexing
